+++ Backup/Communication/DestinationSocketCommunicator.cs
@@ -13,7 +13,7 @@ namespace Communication
     {
         public DestinationSocketCommunicator(string address,
             int port,
-            int bufferSize,
+            long bufferSize,
             ISerialization serialization,
             ILogger logger) : base(address, port, bufferSize, serialization, logger)
         {
